feat: work out the 64-bit arch

+++ jam-pvm/src/test/kotlin/io/forge/jam/pvm/PvmCase.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.Serializable
 data class PvmCase(
     val name: String,
     @SerialName("initial-regs")
-    val initialRegs: List<UInt>,
+    val initialRegs: List<ULong>,
     @SerialName("initial-pc")
     val initialPc: UInt,
     @SerialName("initial-page-map")
